Recording output for readline
Calling readline with an empty prefix does not protect previously output
characters on the same line from being deleted.

Solution: keep track of output saving the last line (after a newline) to
readline_prompt.  Before calling readline, kill the line using an escape
sequence and call it with the saved prompt.

If the start of the line appears too long, it is not likely to be used
as a prompt, so we don’t save it and set a boolean flag.